unittest: properly report import errors
TestLoader's loadTestsFromName method, which is used for finding a test
case to execute, opportunistically tries to import a module in order to
inspect it for tests. If an import error occurs, it strips of the last
part of the name and retries. This logic is required in order to
transparently handle the specification of TestCase names and individual
test methods.
However, this approach is prone to masking errors and potentially
causing wrong imports because it does not differentiate between the case
that a module cannot be found and one where the module was found but
cannot be imported (say, due to a missing dependency). To illustrate the
latter case consider that during development the user added an import
statement to a unittest that cannot be resolved because it has not yet
been defined in the actual product code (a typical scenario in test
driven development). In such a case loadTestsFromName would see an
import error and deduce that the module does not exists. It strips off
the last part of the full name specifier and retries. Now, if another
module would match that name, it would be imported wrongly, for if the
user now goes and implements the import in the product, suddenly a
different module would be imported. If, on the other hand, no module is
found, a generic error is raised that does not detail the actual import
problem.
This change fixes the problem by splitting the import in two parts.
First, we use importlib.util.find_spec (available with Python 3.4) to
check if the given name is a discoverable module. If that is not the
case we strip off a part as before and continue. If a module was found,
however, we perform a normal import and report any error during that
import directly to the user.